# Refactor hb report module