--- a/crm.conf.in
+++ b/crm.conf.in
@@ -56,6 +56,6 @@ ocf_root = @OCF_ROOT_DIR@
 ; from_time = -12H
 ; compress = yes
 ; speed_up = no
-; collect_extra_logs = /var/log/messages /var/log/pacemaker.log
+; collect_extra_logs = /var/log/messages /var/log/ha-cluster-bootstrap.log
 ; remove_exist_dest = no
 ; single_node = no